+++ components/calendar/Calendar.tsx
@@ -1,26 +1,15 @@
 import React from 'react'
 import Subtitle from '@components/Subtitle'
-import listEvents from '@lib/google/calendar'
+import listEvents, { CalendarEvent } from '@lib/google/calendar'
 import InteractiveCalendar from '@components/calendar/InteractiveCalendar'
-
-const CALENDAR_ID = process.env.NEXT_PUBLIC_GOOGLE_CALENDAR_ID
-
-// Compatible with FullCalendar event type
-export type CalendarEvent = {
-  id: string
-  title: string | null | undefined
-  start: string | null | undefined
-  end: string | null | undefined
-  htmlLink: string
-}
 
 const Calendar = async () => {
   const now = new Date()
-  const events = await listEvents(CALENDAR_ID ?? '', now)
+  const events = await listEvents(now)
 
   const fetchMore = async (date: Date): Promise<CalendarEvent[]> => {
     'use server'
-    return listEvents(CALENDAR_ID ?? '', date)
+    return listEvents(date)
   }
 
   return (
